requirements-dev.txt
black
pytest==6.2.5
ipdb==0.13.9
Faker==12.0.0
factory-boy==3.2.1